Add support for finite types

+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/bir/writer/BIRTypeWriter.java
@@ -115,7 +115,7 @@ public class BIRTypeWriter implements TypeVisitor {
     @Override
     public void visit(BFiniteType bFiniteType) {
         buff.writeByte(bFiniteType.tag);
-        buff.writeByte(bFiniteType.valueSpace.size());
+        buff.writeInt(bFiniteType.valueSpace.size());
         for (BLangExpression valueLiteral : bFiniteType.valueSpace) {
             if (!(valueLiteral instanceof BLangLiteral)) {
                 throw new AssertionError(
